Rename qubits module to registers
