+++ src/braket/circuits/ascii_circuit_diagram.py
@@ -23,7 +23,7 @@ from braket.circuits.gate import Gate
 from braket.circuits.instruction import Instruction
 from braket.circuits.noise import Noise
 from braket.circuits.result_type import ResultType
-from braket.qubits.qubit_set import QubitSet
+from braket.registers.qubit_set import QubitSet
 
 
 class AsciiCircuitDiagram(CircuitDiagram):
