+++ src/braket/circuits/gate_calibrations.py
@@ -23,7 +23,7 @@ from braket.circuits.serialization import (
     QubitReferenceType,
 )
 from braket.pulse.pulse_sequence import PulseSequence
-from braket.qubits.qubit_set import QubitSet
+from braket.registers.qubit_set import QubitSet
 
 
 class GateCalibrations:
