+++ src/braket/circuits/noise_model/observable_criteria.py
@@ -21,7 +21,7 @@ from braket.circuits.noise_model.criteria_input_parsing import (
 from braket.circuits.noise_model.result_type_criteria import ResultTypeCriteria
 from braket.circuits.observable import Observable
 from braket.circuits.result_type import ObservableResultType, ResultType
-from braket.qubits.qubit_set import QubitSetInput
+from braket.registers.qubit_set import QubitSetInput
 
 
 class ObservableCriteria(ResultTypeCriteria):
