+++ src/braket/circuits/observables.py
@@ -30,7 +30,7 @@ from braket.circuits.quantum_operator_helpers import (
     verify_quantum_operator_matrix_dimensions,
 )
 from braket.circuits.serialization import IRType, OpenQASMSerializationProperties
-from braket.qubits.qubit_set import QubitSet
+from braket.registers.qubit_set import QubitSet
 
 
 class H(StandardObservable):
