+++ src/braket/circuits/qubit_set.py
@@ -11,4 +11,4 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-from braket.qubits.qubit_set import QubitSet, QubitSetInput  # noqa: F401
+from braket.registers.qubit_set import QubitSet, QubitSetInput  # noqa: F401
