+++ src/braket/registers/__init__.py
@@ -11,5 +11,5 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-from braket.qubits.qubit import Qubit, QubitInput  # noqa: F401
-from braket.qubits.qubit_set import QubitSet, QubitSetInput  # noqa: F401
+from braket.registers.qubit import Qubit, QubitInput  # noqa: F401
+from braket.registers.qubit_set import QubitSet, QubitSetInput  # noqa: F401
