+++ test/unit_tests/braket/registers/test_qubit_set.py
@@ -13,7 +13,7 @@
 
 import pytest
 
-from braket.qubits import Qubit, QubitSet
+from braket.registers import Qubit, QubitSet
 
 
 @pytest.fixture
